Consolidate registration of template filters and globals
Move filters and globals into a single file.

Introduce a single function, setup_jinja_env, for adding all these
to an environment.
This allows setting up the Jinja environment outside of Flask (but
several functions here still depend on the Flask context).
